requirements.txt
django==4.0.3
djangorestframework==3.13.1
django-cors-headers==3.11.0
